Add BaseLogic class Move similar methods from logic classes in  modules to BaseLogic Update scene cleanup Fix bug with deleting segments Add button to restore libraries that changed during Supervisely installation Improve clanup on module reload Update libraries import
